+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.1] - 2022-02-10
+
+### Added
+
+- Fixed Class GetResult
+- Add Class GetLicenseResult for intermediate level between GetResult Class and LicenseGetResult
+- Refactoring of Units Tests
+- Refactoring names of variable in class LicenseGetResult
+
 ## [2.2.0] - 2022-01-04
 
 ### Added